// vue-app/src/shared/interfaces/cart/ICartItem.ts
export interface ICartItem {
  id: number
  cart_id: string
  arrival_date: string
  departure_date: string
  adults: number
  children: number
  units: number
  conservation_fees: number
  southafrican_adults: number
  sadc_adults: number
  international_adults: number
  southafrican_children: number
  sadc_children: number
  international_children: number
  base_pricing: string
  discount_web: any
  community_levy: any
  conservation_fees_amount: any
  camp_id: number
  unit_id: string
  item_type: number
  crs: string
  confirmation: string
  sourcestatus: string
  parent: number
  gmsTimeSlotId: number
}